+++ src-tauri/src/lib.rs
@@ -136,6 +136,7 @@ pub fn run() {
             profile::launch::commands::launch_game,
             profile::launch::commands::get_launch_args,
             profile::launch::commands::open_game_dir,
+            profile::install::commands::install_all_mods,
             profile::install::commands::install_mod,
             profile::install::commands::cancel_all_installs,
             profile::install::commands::has_pending_installations,