+++ src-tauri/src/thunderstore/mod.rs
@@ -167,7 +167,7 @@ impl Thunderstore {
         self.packages_fetched
     }
 
-    /// Returns an iterator over the lastest versions of every package.
+    /// Returns an iterator over the latest versions of every package.
     pub fn latest(&self) -> impl Iterator<Item = BorrowedMod<'_>> {
         self.packages.values().map(move |package| BorrowedMod {
             package,
@@ -178,7 +178,7 @@ impl Thunderstore {
     pub fn get_package(&self, uuid: Uuid) -> Result<&PackageListing> {
         self.packages
             .get(&uuid)
-            .ok_or_else(|| eyre!("package with id {} not found", uuid))
+            .ok_or_else(|| eyre!("package with id {uuid} not found",))
     }
 
     /// Finds a package with the given `full_name` (formatted as `owner-name`).
@@ -186,15 +186,14 @@ impl Thunderstore {
         self.packages
             .values()
             .find(|package| package.ident.as_str() == full_name)
-            .ok_or_else(|| eyre!("package {} not found", full_name))
+            .ok_or_else(|| eyre!("package {full_name} not found",))
     }
 
     pub fn get_mod(&self, package_uuid: Uuid, version_uuid: Uuid) -> Result<BorrowedMod<'_>> {
         let package = self.get_package(package_uuid)?;
         let version = package.get_version(version_uuid).ok_or_else(|| {
             eyre!(
-                "version with id {} not found in package {}",
-                version_uuid,
+                "version with id {version_uuid} not found in package {}",
                 package.ident
             )
         })?;